+++ src/compat/pthread/tests/exit4.c
@@ -114,7 +114,7 @@ static int osThread(void * arg)
   /*
    * Doesn't return and doesn't create an implicit POSIX handle.
    */
-  pthread_exit((void *) result);
+  pthread_exit((void *) (int64_t) result);
 
   return 0;
 }
