+++ src/compat/pthread/tests/once3.c
@@ -98,7 +98,7 @@ mythread(void * arg)
    * eventually cancels only when it becomes the new once thread.
    */
   assert(pthread_cancel(pthread_self()) == 0);
-  assert(pthread_once(&once[(int) arg], myfunc) == 0);
+  assert(pthread_once(&once[(int) (int64_t) arg], myfunc) == 0);
   pte_osMutexLock(numThreads.cs);
   numThreads.i++;
   pte_osMutexUnlock(numThreads.cs);
@@ -127,7 +127,7 @@ int pthread_test_once3()
 
       for (i = 0; i < NUM_THREADS; i++)
         {
-          assert(pthread_create(&t[i][j], NULL, mythread, (void *) j) == 0);
+	    assert(pthread_create(&t[i][j], NULL, mythread, (void *) (int64_t) j) == 0);
         }
     }
